Still not sure about balance thing.

+++ 9-binary_tree_height.c
@@ -2,11 +2,10 @@
 
 /**
  * measure_branches - measure the subbranches of the tree
- * @t: node of the root or sub-root of the tree
- * @l: place to store the height down the left
- * @r: place to store the height down the right
+ * @tree: node of the root or sub-root of the tree
+ * @counters: pointer to counter array
 */
-static void measure_branches(const binary_tree_t *tree, int *counters)
+static void measure_branches(const binary_tree_t *tree, size_t *counters)
 {
 	counters[LEFT] = (
 		(tree->left) ? 1 + binary_tree_height(tree->left) : 0
@@ -22,14 +21,14 @@ static void measure_branches(const binary_tree_t *tree, int *counters)
  *
  * Return: Tree is NULL, function returns 0, else returns the height.
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int counters[2];
+	size_t counters[2];
 
 	if (tree == NULL)
 		return (0);
 
-	measure_branches(tree, (int *)&counters);
+	measure_branches(tree, (size_t *)&counters);
 
 	return (
 		(counters[LEFT] > counters[RIGHT]) ? counters[LEFT] : counters[RIGHT]
